Disable maps with no_plugin_maps or no_xml_maps
There's no option to define all of your own maps, so we should allow
them all to be turned off.

An alternative solution would be to map LocalLeader to something bizarre
that you won't type and set g:xml_tag_completion_map to something
beginning with LocalLeader.

+++ ftplugin/xml.vim
@@ -544,37 +544,41 @@ endif
 setlocal matchpairs+=<:>
 setlocal commentstring=<!--%s-->
 
-" Have this as an escape incase you want a literal '>' not to run the
-" ParseTag function.
-if !exists("g:xml_tag_completion_map")
-    inoremap <buffer> <LocalLeader>. >
-    inoremap <buffer> <LocalLeader>> >
+if (! exists('no_plugin_maps') || ! no_plugin_maps) &&
+      \ (! exists('g:no_xml_maps') || ! g:no_xml_maps)
+
+    " Have this as an escape incase you want a literal '>' not to run the
+    " ParseTag function.
+    if !exists("g:xml_tag_completion_map")
+        inoremap <buffer> <LocalLeader>. >
+        inoremap <buffer> <LocalLeader>> >
+    endif
+
+    " Jump between the beggining and end tags.
+    nnoremap <buffer> <LocalLeader>5 :call <SID>TagMatch1()<Cr>
+    nnoremap <buffer> <LocalLeader>% :call <SID>TagMatch1()<Cr>
+    vnoremap <buffer> <LocalLeader>5 <Esc>:call <SID>VisualTag()<Cr>
+    vnoremap <buffer> <LocalLeader>% <Esc>:call <SID>VisualTag()<Cr>
+
+    " Wrap selection in XML tag
+    vnoremap <buffer> <LocalLeader>x "xx:call <SID>WrapTag(@x)<Cr>
+    nnoremap <buffer> <LocalLeader>d :call <SID>DeleteTag()<Cr>
+
+    " Parse the tag after pressing the close '>'.
+    if !exists("g:xml_tag_completion_map")
+        " inoremap <buffer> > ><Esc>:call <SID>ParseTag()<Cr>
+        inoremap <buffer> > <Esc>:call <SID>InsertGt()<Cr>
+    else
+        execute "inoremap <buffer> " . g:xml_tag_completion_map . " <Esc>:call <SID>InsertGt()<Cr>"
+    endif
+
+    nnoremap <buffer> <LocalLeader><LocalLeader> :call <SID>EditFromJump()<Cr>
+    inoremap <buffer> <LocalLeader><LocalLeader> <Esc>:call <SID>EditFromJump()<Cr>
+    " Clear out all left over xml_jump_string garbage
+    nnoremap <buffer> <LocalLeader>w :call <SID>ClearJumpMarks()<Cr>
+    " The syntax files clear out any predefined syntax definitions. Recreate
+    " this when ever a xml_jump_string is created. (in ParseTag)
 endif
-
-" Jump between the beggining and end tags.
-nnoremap <buffer> <LocalLeader>5 :call <SID>TagMatch1()<Cr>
-nnoremap <buffer> <LocalLeader>% :call <SID>TagMatch1()<Cr>
-vnoremap <buffer> <LocalLeader>5 <Esc>:call <SID>VisualTag()<Cr>
-vnoremap <buffer> <LocalLeader>% <Esc>:call <SID>VisualTag()<Cr>
-
-" Wrap selection in XML tag
-vnoremap <buffer> <LocalLeader>x "xx:call <SID>WrapTag(@x)<Cr>
-nnoremap <buffer> <LocalLeader>d :call <SID>DeleteTag()<Cr>
-
-" Parse the tag after pressing the close '>'.
-if !exists("g:xml_tag_completion_map")
-    " inoremap <buffer> > ><Esc>:call <SID>ParseTag()<Cr>
-    inoremap <buffer> > <Esc>:call <SID>InsertGt()<Cr>
-else
-    execute "inoremap <buffer> " . g:xml_tag_completion_map . " <Esc>:call <SID>InsertGt()<Cr>"
-endif
-
-nnoremap <buffer> <LocalLeader><LocalLeader> :call <SID>EditFromJump()<Cr>
-inoremap <buffer> <LocalLeader><LocalLeader> <Esc>:call <SID>EditFromJump()<Cr>
-" Clear out all left over xml_jump_string garbage
-nnoremap <buffer> <LocalLeader>w :call <SID>ClearJumpMarks()<Cr>
-" The syntax files clear out any predefined syntax definitions. Recreate
-" this when ever a xml_jump_string is created. (in ParseTag)
 
 augroup xml
     au!
